+++ tomcat/src/main/java/nextstep/jwp/service/UserService.java
@@ -1,11 +1,8 @@
 package nextstep.jwp.service;
 
 import java.util.Map;
-import java.util.UUID;
 import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.model.User;
-import org.apache.catalina.Session;
-import org.apache.catalina.SessionManager;
 import org.apache.coyote.exception.UnauthorizedException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -13,7 +10,7 @@ import org.slf4j.LoggerFactory;
 public class UserService {
     private static final Logger log = LoggerFactory.getLogger(UserService.class);
 
-    public String logIn(final Map<String, String> requestBody) {
+    public User logIn(final Map<String, String> requestBody) {
         final String account = requestBody.get("account");
         final String password = requestBody.get("password");
 
@@ -25,18 +22,10 @@ public class UserService {
         }
         log.info("{}", user);
 
-        return setSession(user);
+        return user;
     }
 
-    private String setSession(final User user) {
-        final String id = UUID.randomUUID().toString();
-        final Session session = new Session(id);
-        session.setAttribute("user", user);
-        SessionManager.add(session);
-        return id;
-    }
-
-    public String register(final Map<String, String> requestBody) {
+    public User register(final Map<String, String> requestBody) {
         final String account = requestBody.get("account");
         final String password = requestBody.get("password");
         final String email = requestBody.get("email");
@@ -49,7 +38,7 @@ public class UserService {
         final User user = new User(account, password, email);
         InMemoryUserRepository.save(user);
 
-        return setSession(user);
+        return user;
     }
 
 }
